backend/requirements.txt
# Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Background Jobs
arq==0.25.0
croniter==2.0.1

# Security
cryptography==42.0.0

# Vector Database
chromadb==0.4.22

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# AI/ML
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Type Checking
mypy==1.8.0
types-python-dateutil==2.8.19.20240106
